Fix issues with args
